made exp_global optional, updated serializer accordingly; added PID and EXP_DIR expressions
