+++ xnmt/exp_global.py
@@ -8,9 +8,10 @@ from xnmt.serialize.serializable import Serializable
 
 class ExpGlobal(Serializable):
   yaml_tag = u'!ExpGlobal'
-  def __init__(self, model_file=None, out_file=None, dropout = 0.0,
-               weight_noise = 0.0, default_layer_dim = 512, save_num_checkpoints=1,
-               eval_only=False, commandline_args=None,
+  def __init__(self, model_file="{EXP_DIR}/models/{EXP}.mod", 
+               out_file="{EXP_DIR}/models/{EXP}.mod",
+               dropout = 0.0, weight_noise = 0.0, default_layer_dim = 512,
+               save_num_checkpoints=1, eval_only=False, commandline_args=None,
                dynet_param_collection = None):
     self.model_file = model_file
     self.out_file = out_file
